Refactor: Migrate CSS to SCSS modules and update Storybook config
This commit replaces global CSS files with SCSS modules for better component styling. It also updates the Storybook configuration to use the new module structure and refines the test command in package.json.
