+++ design-system/.storybook/preview.js
@@ -1,30 +1,2 @@
-import React from 'react';
-import ThemeProvider from '../src/theme/ThemeProvider.jsx';
-
-/** @type { import('@storybook/react-vite').Preview } */
-const preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-    a11y: {
-      test: "todo"
-    }
-  },
-  decorators: [
-    (Story) => {
-      return (
-        <ThemeProvider>
-          <div id="storybook-root">
-            <Story />
-          </div>
-        </ThemeProvider>
-      );
-    }
-  ]
-};
-
-export default preview;
+// Moved to preview.jsx to allow JSX syntax
+export { default } from './preview.jsx';
