+++ design-system/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview",
     "storybook": "storybook dev -p 6006 --quiet",
     "build-storybook": "storybook build",
-    "test": "vitest run --reporter=verbose",
+    "test": "vitest run --project=unit --reporter=verbose",
     "test:ui": "vitest --ui --environment=jsdom",
     "test:sb": "vitest --project=storybook --run"
   },
