+++ design-system/src/App.module.scss
@@ -14,7 +14,7 @@
 .logo:hover {
   filter: drop-shadow(0 0 2em #646cffaa);
 }
-.logo.react:hover {
+.react:hover {
   filter: drop-shadow(0 0 2em #61dafbaa);
 }
 
@@ -37,6 +37,6 @@
   padding: 2em;
 }
 
-.read-the-docs {
+.readTheDocs {
   color: #888;
 }
