+++ design-system/src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+// Global CSS removed in favor of theme variables and SCSS modules
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
