+++ design-system/src/stories/Header.jsx
@@ -3,11 +3,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { Button } from './Button';
-import './header.css';
+import cls from './header.module.scss';
 
 export const Header = ({ user = null, onLogin, onLogout, onCreateAccount }) => (
   <header>
-    <div className="storybook-header">
+    <div className={cls.storybookHeader}>
       <div>
         <svg width="32" height="32" viewBox="0 0 32 32" xmlns="http://www.w3.org/2000/svg">
           <g fill="none" fillRule="evenodd">
@@ -30,7 +30,7 @@ export const Header = ({ user = null, onLogin, onLogout, onCreateAccount }) => (
       <div>
         {user ? (
           <>
-            <span className="welcome">
+            <span className={cls.welcome}>
               Welcome, <b>{user.name}</b>!
             </span>
             <Button size="small" onClick={onLogout} label="Log out" />
