+++ design-system/src/stories/button.module.scss
@@ -1,0 +1,14 @@
+.storybookButton {
+  display: inline-block;
+  cursor: pointer;
+  border: 0;
+  border-radius: 3em;
+  font-weight: 700;
+  line-height: 1;
+  font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+}
+.primary { background-color: #555ab9; color: white; }
+.secondary { box-shadow: rgba(0, 0, 0, 0.15) 0 0 0 1px inset; background-color: transparent; color: #333; }
+.small { padding: 10px 16px; font-size: 12px; }
+.medium { padding: 11px 20px; font-size: 14px; }
+.large { padding: 12px 24px; font-size: 16px; }
