+++ design-system/src/stories/page.module.scss
@@ -1,0 +1,28 @@
+.storybookPage {
+  margin: 0 auto;
+  padding: 48px 20px;
+  max-width: 600px;
+  color: #333;
+  font-size: 14px;
+  line-height: 24px;
+  font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+}
+
+.storybookPage h2 {
+  display: inline-block;
+  vertical-align: top;
+  margin: 0 0 4px;
+  font-weight: 700;
+  font-size: 32px;
+  line-height: 1;
+}
+
+.storybookPage p { margin: 1em 0; }
+.storybookPage a { color: inherit; }
+.storybookPage ul { margin: 1em 0; padding-left: 30px; }
+.storybookPage li { margin-bottom: 8px; }
+
+.tip { display: inline-block; vertical-align: top; margin-right: 10px; border-radius: 1em; background: #e7fdd8; padding: 4px 12px; color: #357a14; font-weight: 700; font-size: 11px; line-height: 12px; }
+.tipWrapper { margin-top: 40px; margin-bottom: 40px; font-size: 13px; line-height: 20px; }
+.tipWrapper svg { display: inline-block; vertical-align: top; margin-top: 3px; margin-right: 4px; width: 12px; height: 12px; }
+.tipWrapper svg path { fill: #1ea7fd; }
